[REFACTOR] change temporary file name
