+++ cmd/utils/logger.go
@@ -1,35 +1,35 @@
 package utils
 
-import (
-	"fmt"
-	"log"
-	"os"
-	"path/filepath"
-)
+type AdvancedLogger interface {
+	Info(message string)
+	Debug(message string)
+	Error(message string)
+	Panic(message string)
+	Fatal(message string)
+}
 
-var GeneralLogger *log.Logger
-var ErrorLogger *log.Logger
-var DebugLogger *log.Logger
-var InfoLogger *log.Logger
-var PanicLogger *log.Logger
-var FatalLogger *log.Logger
+type ConsoleLogger struct{}
 
-// TODO
-// Use builder pattern to create a more descriptive logging functions
-func init() {
-	absPath, err := filepath.Abs(`\ppl\kompas-gramedia\be\log`)
-	if err != nil {
-		fmt.Println("Error reading abs path: ", err)
-	}
-	generalLog, err := os.OpenFile(absPath+"/general-log.log", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
-	if err != nil {
-		fmt.Println("error opening file: ", err)
-		os.Exit(1)
-	}
-	GeneralLogger = log.New(generalLog, "General Logger:\t", log.Ldate|log.Ltime|log.Lshortfile)
-	ErrorLogger = log.New(generalLog, "Error:\t", log.Ldate|log.Ltime|log.Lshortfile)
-	DebugLogger = log.New(generalLog, "Debug:\t", log.Ldate|log.Ltime|log.Lshortfile)
-	InfoLogger = log.New(generalLog, "Info:\t", log.Ldate|log.Ltime|log.Lshortfile)
-	PanicLogger = log.New(generalLog, "Panic:\t", log.Ldate|log.Ltime|log.Lshortfile)
-	FatalLogger = log.New(generalLog, "Fatal:\t", log.Ldate|log.Ltime|log.Lshortfile)
+func (c ConsoleLogger) Info(message string) {
+	InfoLogger.Println(message)
+}
+
+func (c ConsoleLogger) Debug(message string) {
+	DebugLogger.Println(message)
+}
+
+func (c ConsoleLogger) Error(message string) {
+	ErrorLogger.Println(message)
+}
+
+func (c ConsoleLogger) Panic(message string) {
+	PanicLogger.Panicln(message)
+}
+
+func (c ConsoleLogger) Fatal(message string) {
+	FatalLogger.Fatalln(message)
+}
+
+func NewConsoleLogger() *ConsoleLogger {
+	return &ConsoleLogger{}
 }
